assets: split out DCS object handling from StaticAsset
This allows the same DCS object handling to be reused across other
assets that need to deal with DCS objects.
